+++ sshd-core/src/main/java/org/apache/sshd/server/subsystem/SubsystemFactory.java
@@ -19,14 +19,46 @@
 
 package org.apache.sshd.server.subsystem;
 
-import org.apache.sshd.common.NamedFactory;
+import java.io.IOException;
+import java.util.Collection;
+
+import org.apache.sshd.common.NamedResource;
+import org.apache.sshd.server.channel.ChannelSession;
 import org.apache.sshd.server.command.Command;
 
 /**
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
-// CHECKSTYLE:OFF
-public interface SubsystemFactory extends NamedFactory<Command> {
-    // nothing extra
+public interface SubsystemFactory extends NamedResource {
+    /**
+     * @param channel The {@link ChannelSession} through which the command
+     * has been received
+     * @return a non {@code null} {@link Command} instance representing
+     * the subsystem to be run
+     * @throws IOException if failed to create the instance
+     */
+    Command createSubsystem(ChannelSession channel) throws IOException;
+
+    /**
+     * @param channel The {@link ChannelSession} through which the command
+     * has been received
+     * @param factories The available {@link SubsystemFactory}-ies - ignored
+     * if {@code null}/empty
+     * @param name Requested subsystem name
+     * @return The created {@link Command} instance representing
+     * the subsystem to be run - {@code null} if no match found
+     * @throws IOException If found a matching factory but failed to create
+     * the command instance
+     */
+    static Command createSubsystem(
+            ChannelSession channel, Collection<? extends SubsystemFactory> factories, String name)
+                throws IOException {
+        SubsystemFactory factory =
+            NamedResource.findByName(name, String.CASE_INSENSITIVE_ORDER, factories);
+        if (factory != null) {
+            return factory.createSubsystem(channel);
+        } else {
+            return null;
+        }
+    }
 }
-//CHECKSTYLE:ON